Move AstIdMap infra to the span crate
