+++ crates/hir-def/src/nameres.rs
@@ -61,13 +61,13 @@ use std::ops::Deref;
 
 use base_db::{CrateId, Edition, FileId};
 use hir_expand::{
-    ast_id_map::FileAstId, name::Name, proc_macro::ProcMacroKind, HirFileId, InFile, MacroCallId,
-    MacroDefId,
+    name::Name, proc_macro::ProcMacroKind, HirFileId, InFile, MacroCallId, MacroDefId,
 };
 use itertools::Itertools;
 use la_arena::Arena;
 use profile::Count;
 use rustc_hash::{FxHashMap, FxHashSet};
+use span::FileAstId;
 use stdx::format_to;
 use syntax::{ast, SmolStr};
 use triomphe::Arc;
